killed Direct Line v1 support. Direct Line v3 is now default and only protocol supported
