+++ public/botchat.d.ts
@@ -6,14 +6,12 @@ export interface SecretOrToken {
 }
 
 export declare class DirectLine {
-    constructor(secretOrToken: SecretOrToken, domain?: string);
-}
-
-export declare class DirectLine3 {
-    constructor(secretOrToken: SecretOrToken, domain?: string, segment?: string);
+    constructor(secretOrToken: SecretOrToken, domain?: string, segment?: string); // segment is DEPRECATED and will be removed before release
     start();
     end();
 }
+
+export type DirectLine3 = DirectLine; // DEPRECATED will be removed before release 
 
 export interface FormatOptions {
     showHeader?: boolean
