+++ src/BotChat.ts
@@ -1,5 +1,4 @@
 export { App, AppProps } from './App';
 export { Chat, ChatProps, FormatOptions } from './Chat';
-export { SecretOrToken, DirectLine } from './directLine';
-export { DirectLine3 } from './directLine3';
+export { SecretOrToken, DirectLine, DirectLine as DirectLine3 } from './directLine';
 export { IBotConnection } from './BotConnection';
